# docs/api/openapi.yaml
...
openapi: "3.0.3"
info:
  title: "particeps Agent API"
  version: "v1"
  description: "对应当前 Agent 注册的全部 HTTP 操作。会话管理操作须通过同源检查；代理应保留原 Host。默认启用 Secure Cookie，实验明文 HTTP 需显式设置 session_cookie_secure: false。所有 JSON 响应禁止缓存。"
servers:
  -
    url: "/api/v1"
...
security:
  -
    sessionCookie: []
  -
    bearerToken: []
paths:
  "/session":
    post:
      operationId: "login"
      summary: "管理员登录"
      tags:
        - "认证"
      description: "每个连接来源每分钟最多 5 次尝试，全局每分钟最多 30 次，最多同时运行 2 次密码校验。忽略外部提供的来源地址头；反向代理部署还可在代理层按客户端限速。"
      security: []
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: "成功。"
          headers:
            Set-Cookie:
              description: "particeps_session；HttpOnly、SameSite=Strict；默认 Secure，12 小时后到期。"
              schema:
                type: "string"
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OK"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
    delete:
      operationId: "logout"
      summary: "撤销当前浏览器会话"
      tags:
        - "认证"
      description: "通过来源检查后撤销 Cookie 指定的会话。没有会话也可退出；撤销失败返回 500，并保留 Cookie 供重试。此接口不撤销 Bearer Token。"
      security: []
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OK"
        "403":
          "$ref": "#/components/responses/Error403"
        "500":
          "$ref": "#/components/responses/Error500"
  "/meta":
    get:
      operationId: "getMeta"
      summary: "产品与 API 版本"
      tags:
        - "管理"
      x-permissions:
        - "admin"
        - "read"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Meta"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/host":
    get:
      operationId: "getHost"
      summary: "母机资源概览"
      tags:
        - "管理"
      x-permissions:
        - "admin"
        - "read"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Host"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/host/network":
    get:
      operationId: "detectNetwork"
      summary: "探测母机地址候选"
      tags:
        - "管理"
      x-permissions:
        - "admin"
        - "read"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Network"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/settings/network-pool":
    get:
      operationId: "getNetworkPool"
      summary: "读取地址池设置"
      tags:
        - "管理"
      x-permissions:
        - "admin"
        - "read"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Pool"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
    put:
      operationId: "saveNetworkPool"
      summary: "保存地址池设置"
      tags:
        - "管理"
      description: "当前保存设置；不承诺完成地址归属校验、指定出口、独立 IPv4、IPv6/NAT66 或外部可达性验证。"
      x-permissions:
        - "admin"
        - "manage"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/Pool"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Pool"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/settings/cpu-cap":
    put:
      operationId: "setCPUCap"
      summary: "修改小鸡合计 CPU 硬上限"
      tags:
        - "管理"
      description: "不能低于最大单台额度；失败时需查看回读状态，不能将请求受理视为内核设置成功。"
      x-permissions:
        - "admin"
        - "manage"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/CapRequest"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/CapResult"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/instances":
    get:
      operationId: "listInstances"
      summary: "列出受管实例"
      tags:
        - "管理"
      x-permissions:
        - "admin"
        - "read"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Instances"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "500":
          "$ref": "#/components/responses/Error500"
    post:
      operationId: "createInstances"
      summary: "提交批量创建任务"
      tags:
        - "管理"
      description: "202 只表示任务受理，须查询逐实例结果。密码不出现在普通任务查询中。"
      x-permissions:
        - "admin"
        - "manage"
      parameters:
        -
          name: "Idempotency-Key"
          in: "header"
          schema:
            type: "string"
          description: "相同键与相同请求返回原任务；相同键与不同请求拒绝。"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/CreateRequest"
      responses:
        "202":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Task"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "409":
          "$ref": "#/components/responses/Error409"
  "/instances/{id}":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
    get:
      operationId: "getInstance"
      summary: "查询实例与端口状态"
      tags:
        - "管理"
      x-permissions:
        - "admin"
        - "read"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Instance"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
  "/instances/{id}/start":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
    post:
      operationId: "startInstance"
      summary: "启动实例"
      tags:
        - "管理"
      description: "同实例的电源、删除和密码操作串行。进行中的创建会拒绝操作；未知后端结果保留待核对状态。"
      x-permissions:
        - "admin"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OK"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/instances/{id}/stop":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
    post:
      operationId: "stopInstance"
      summary: "优雅停止实例"
      tags:
        - "管理"
      description: "同实例的电源、删除和密码操作串行。进行中的创建会拒绝操作；未知后端结果保留待核对状态。"
      x-permissions:
        - "admin"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OK"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/instances/{id}/force-stop":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
    post:
      operationId: "forcestopInstance"
      summary: "强制停止实例"
      tags:
        - "管理"
      description: "同实例的电源、删除和密码操作串行。进行中的创建会拒绝操作；未知后端结果保留待核对状态。"
      x-permissions:
        - "admin"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OK"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/instances/{id}/restart":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
    post:
      operationId: "restartInstance"
      summary: "重启实例"
      tags:
        - "管理"
      description: "同实例的电源、删除和密码操作串行。进行中的创建会拒绝操作；未知后端结果保留待核对状态。"
      x-permissions:
        - "admin"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OK"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/instances/{id}/delete":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
    post:
      operationId: "deleteInstance"
      summary: "删除实例并清理所属资源"
      tags:
        - "管理"
      description: "清理失败保留占用。成功删除同时使该实例尚未领取的初始凭据失效。"
      x-permissions:
        - "admin"
        - "manage"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OK"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/instances/{id}/rebuild":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
    post:
      operationId: "rebuildInstance"
      summary: "重装接口占位"
      tags:
        - "管理"
      description: "当前未实现；返回 400，不停止实例、不更换系统盘。"
      x-permissions:
        - "admin"
        - "manage"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/RebuildRequest"
      responses:
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/instances/{id}/resources":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
    patch:
      operationId: "patchResources"
      summary: "修改资源配置"
      tags:
        - "管理"
      description: "系统盘只允许增大。结果不明时保留待核对状态，须重新读取 resourceStatus。"
      x-permissions:
        - "admin"
        - "manage"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/ResourceRequest"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/OK"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
  "/instances/{id}/ports":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
    post:
      operationId: "addPort"
      summary: "追加一对 TCP/UDP 号码"
      tags:
        - "管理"
      description: "号码为 0 或省略时自动分配；两协议同归属。409 时可能已保存分配，须读取 networkStatus。"
      x-permissions:
        - "admin"
        - "manage"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/PortAdd"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Instance"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "409":
          "$ref": "#/components/responses/Error409"
  "/instances/{id}/ports/{number}/{proto}":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
        description: "管理 ID 或实例名称；其他路径使用对应对象 ID。"
      -
        name: "number"
        in: "path"
        required: true
        schema:
          type: "integer"
          minimum: 1
          maximum: 65535
      -
        name: "proto"
        in: "path"
        required: true
        schema:
          type: "string"
          enum:
            - "tcp"
            - "udp"
    patch:
      operationId: "editPort"
      summary: "修改协议规则的目标端口"
      tags:
        - "管理"
      description: "保留号码归属并清理旧目标连接。失败可能已保存目标，须重新查询，不依赖旧页面缓存。"
      x-permissions:
        - "admin"
        - "manage"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/PortEdit"
      responses:
        "200":
          description: "成功。"
          headers:
            Cache-Control:
              schema:
                type: "string"
                enum:
                  - "no-store"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Instance"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "409":
          "$ref": "#/components/responses/Error409"
...
components:
  securitySchemes:
    sessionCookie:
      type: "apiKey"
      in: "cookie"
      name: "particeps_session"
    bearerToken:
      type: "http"
      scheme: "bearer"
      description: "独立只读或管理 Token；角色由 Agent 数据库校验。"
  schemas:
...
    OK:
      type: "object"
      required:
        - "ok"
      properties:
        ok:
          type: "string"
          enum:
            - "1"
    Meta:
      type: "object"
      required:
        - "product"
        - "api"
      properties:
        product:
          type: "string"
          enum:
            - "particeps"
        api:
          type: "string"
          enum:
            - "v1"
    LoginRequest:
      type: "object"
      required:
        - "password"
      properties:
        password:
          type: "string"
          format: "password"
      additionalProperties: false
...
    Network:
      type: "object"
      required:
        - "addresses"
      properties:
        addresses:
          type: "array"
          items:
            "$ref": "#/components/schemas/Address"
...
  responses:
    Error400:
      description: "请求字段、参数或当前状态不允许此操作。"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    Error401:
      description: "未认证或凭据无效。"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    Error403:
      description: "权限不足或会话来源校验失败。"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
...
    Error429:
      description: "登录尝试达到限制，请在 Retry-After 指定的秒数后重试。"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
      headers:
        Retry-After:
          description: "至少等待的秒数。"
          schema:
            type: "integer"
            minimum: 1
    Error500:
      description: "服务或存储失败；不能视为操作成功。"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
...